优化接口请求的参数，重构路由

+++ src/main/kotlin/com/hgm/routes/CommentRoutes.kt
@@ -8,7 +8,7 @@ import com.hgm.service.LikeService
 import com.hgm.service.UserService
 import com.hgm.utils.ApiResponseMessage
 import com.hgm.utils.QueryParams
-import com.hgm.utils.ifEmailBelongsToUser
+import com.hgm.utils.userId
 import io.ktor.application.*
 import io.ktor.auth.*
 import io.ktor.http.*
@@ -18,7 +18,6 @@ import io.ktor.routing.*
 
 fun Route.addComment(
     commentService: CommentService,
-    userService: UserService
 ) {
     authenticate {
         post("/api/comment/add") {
@@ -27,40 +26,35 @@ fun Route.addComment(
                 return@post
             }
 
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                onValidateEmail = userService::doesEmailBelongToUserId
-            ) {
-                when (commentService.addComment(request)) {
-                    CommentService.ValidationEvent.CommentTooLong -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BaseResponse(
-                                successful = false,
-                                message = ApiResponseMessage.COMMENT_LENGTH_TOO_LONG
-                            )
+            when (commentService.addComment(request, call.userId)) {
+                CommentService.ValidationEvent.CommentTooLong -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BaseResponse(
+                            successful = false,
+                            message = ApiResponseMessage.COMMENT_LENGTH_TOO_LONG
                         )
-                    }
+                    )
+                }
 
-                    CommentService.ValidationEvent.FieldEmpty -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BaseResponse(
-                                successful = false,
-                                message = ApiResponseMessage.FIELDS_BLANK
-                            )
+                CommentService.ValidationEvent.FieldEmpty -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BaseResponse(
+                            successful = false,
+                            message = ApiResponseMessage.FIELDS_BLANK
                         )
-                    }
+                    )
+                }
 
-                    CommentService.ValidationEvent.Success -> {
-                        call.respond(
-                            HttpStatusCode.OK,
-                            BaseResponse(
-                                successful = true,
-                                message = ApiResponseMessage.ADD_COMMENT_SUCCESSFUL
-                            )
+                CommentService.ValidationEvent.Success -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BaseResponse(
+                            successful = true,
+                            message = ApiResponseMessage.ADD_COMMENT_SUCCESSFUL
                         )
-                    }
+                    )
                 }
             }
         }
@@ -78,7 +72,7 @@ fun Route.getCommentByPost(
                 return@get
             }
             val comments = commentService.getCommentByPost(postId)
-            call.respond(HttpStatusCode.OK,comments)
+            call.respond(HttpStatusCode.OK, comments)
         }
     }
 }
@@ -86,38 +80,41 @@ fun Route.getCommentByPost(
 
 fun Route.deleteComment(
     commentService: CommentService,
-    userService: UserService,
     likeService: LikeService
-){
+) {
     authenticate {
-        delete("/api/comment/delete"){
-            val request=call.receiveOrNull<DeleteCommentRequest>()?: kotlin.run {
+        delete("/api/comment/delete") {
+            val request = call.receiveOrNull<DeleteCommentRequest>() ?: kotlin.run {
                 call.respond(HttpStatusCode.BadRequest)
                 return@delete
             }
 
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                onValidateEmail = userService::doesEmailBelongToUserId
-            ){
-                val deleteSuccessful = commentService.deleteComment(request.commentId)
-                if (deleteSuccessful){
-                    likeService.removeLikeByParent(request.commentId)
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BaseResponse(
-                            successful = true,
-                            message = ApiResponseMessage.DELETE_COMMENT_SUCCESSFUL
-                        )
+            val comment = commentService.getCommentById(request.commentId) ?: kotlin.run {
+                call.respond(HttpStatusCode.BadRequest)
+                return@delete
+            }
+            if (comment.userId != call.userId) {
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
+            }
+
+            val deleteSuccessful = commentService.deleteComment(request.commentId)
+            if (deleteSuccessful) {
+                likeService.removeLikeByParent(request.commentId)
+                call.respond(
+                    HttpStatusCode.OK,
+                    BaseResponse(
+                        successful = true,
+                        message = ApiResponseMessage.DELETE_COMMENT_SUCCESSFUL
                     )
-                }else{
-                    call.respond(
-                        HttpStatusCode.NotFound,
-                        BaseResponse(
-                            successful = false,
-                        )
+                )
+            } else {
+                call.respond(
+                    HttpStatusCode.NotFound,
+                    BaseResponse(
+                        successful = false,
                     )
-                }
+                )
             }
         }
     }
